Old version & new version karaoke management

+++ projects/ql_karaoke/src/main/java/info/kinhho/karaoke/services/RoomPriceService.java
@@ -17,6 +17,8 @@ public class RoomPriceService {
 
 	public void createSeedData() {
 		
+		
+		
 	}
 
 	public Iterable<RoomPrice> get() {
@@ -26,7 +28,7 @@ public class RoomPriceService {
 	
 	public RoomPrice get(String id) {
 		
-		return roomPriceRepository.findById(Long.parseLong(id)).get();
+		return roomPriceRepository.findById(Integer.parseInt(id)).get();
 	}
